Better test automation with cmsis build tools
Added a python script for automation of tests with cmsis build tools.
Correction to some tests
Float promotion issues in some f16 code.

+++ Source/SupportFunctions/arm_f16_to_q15.c
@@ -124,8 +124,8 @@ void arm_f16_to_q15(
          * convert from float to Q31 and then store the results in the destination buffer
          */
         in = *pIn++;
-        in = (in * 32768.0);
-        in += in > 0.0 ? 0.5 : -0.5;
+        in = ((_Float16)in * (_Float16)32768.0f16);
+        in += (_Float16)in > 0.0f16 ? 0.5f16 : -0.5f16;
         *pDst++ = clip_q31_to_q15((q31_t) (in));
 
 #else
